6.4.0

+++ rsscrawler/ombi.py
@@ -17,12 +17,12 @@ from rsscrawler.url import post_url_json
 def mdb(configfile, dbfile, tmdbid, mdb_api, log_debug):
     get_title = get_url_headers(
         'https://api.themoviedb.org/3/movie/' + str(tmdbid) + '?api_key=' + mdb_api + '&language=de-DE', configfile,
-        dbfile, headers={'Content-Type': 'application/json'})
+        dbfile, headers={'Content-Type': 'application/json'})[0]
     raw_title = json.loads(get_title.text).get("title")
     if not raw_title:
         get_title = get_url_headers(
             'https://api.themoviedb.org/3/movie/' + str(tmdbid) + '?api_key=' + mdb_api + '&language=en-US', configfile,
-            dbfile, headers={'Content-Type': 'application/json'})
+            dbfile, headers={'Content-Type': 'application/json'})[0]
         raw_title = json.loads(get_title.text).get("title")
     if raw_title:
         title = sanitize(raw_title)
@@ -39,13 +39,14 @@ def get_tvdb_token(configfile, dbfile, tvd_user, tvd_userkey, tvd_api, log_debug
         'userkey': tvd_userkey,
         'apikey': tvd_api,
     })
-    response = json.loads(response)
-    token = response.get('token')
-    db.delete("tvdb_token")
-    db.store("tvdb_token", token)
+    if response:
+        response = json.loads(response)
+        token = response.get('token')
+        db.delete("tvdb_token")
+        db.store("tvdb_token", token)
 
-    if token:
-        return token
+        if token:
+            return token
     else:
         log_debug("Aufgrund fehlerhafter API-Zugangsdaten werden keine Serien aus Ombi importiert.")
         return False
@@ -61,7 +62,7 @@ def tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug):
     if token:
         get_info = get_url_headers('https://api.thetvdb.com/series/' + str(tvdbid), configfile, dbfile,
                                    headers={'Authorization': 'Bearer ' + token, 'Content-Type': 'application/json',
-                                            'Accept': 'application/json', 'Accept-Language': 'de'})
+                                            'Accept': 'application/json', 'Accept-Language': 'de'})[0]
 
         if get_info.status_code == 401:
             token = get_tvdb_token(configfile, dbfile, tvd_user, tvd_userkey, tvd_api, log_debug)
@@ -69,7 +70,7 @@ def tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug):
                 get_info = get_url_headers('https://api.thetvdb.com/series/' + str(tvdbid), configfile, dbfile,
                                            headers={'Authorization': 'Bearer ' + token,
                                                     'Content-Type': 'application/json',
-                                                    'Accept': 'application/json', 'Accept-Language': 'de'})
+                                                    'Accept': 'application/json', 'Accept-Language': 'de'})[0]
             else:
                 return False
 
@@ -79,7 +80,7 @@ def tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug):
         if not raw_title:
             get_info = get_url_headers('https://api.thetvdb.com/series/' + str(tvdbid), configfile, dbfile,
                                        headers={'Authorization': 'Bearer ' + token, 'Content-Type': 'application/json',
-                                                'Accept': 'application/json', 'Accept-Language': 'en'})
+                                                'Accept': 'application/json', 'Accept-Language': 'en'})[0]
             raw_data = json.loads(get_info.text)
             raw_info = raw_data.get('data')
             raw_title = raw_info.get('seriesName')
@@ -87,7 +88,7 @@ def tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug):
         get_episodes = get_url_headers('https://api.thetvdb.com/series/' + str(tvdbid) + '/episodes', configfile,
                                        dbfile,
                                        headers={'Authorization': 'Bearer ' + token, 'Content-Type': 'application/json',
-                                                'Accept': 'application/json', 'Accept-Language': 'de'})
+                                                'Accept': 'application/json', 'Accept-Language': 'de'})[0]
         raw_episode_data = json.loads(get_episodes.text)
         episodes = raw_episode_data.get('data')
         total_pages = raw_episode_data.get('links')
@@ -101,7 +102,7 @@ def tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug):
                         dbfile,
                         headers={'Authorization': 'Bearer ' + token,
                                  'Content-Type': 'application/json',
-                                 'Accept': 'application/json', 'Accept-Language': 'de'})
+                                 'Accept': 'application/json', 'Accept-Language': 'de'})[0]
                     raw_episode_data = json.loads(get_episodes.text)
                     more_episodes = raw_episode_data.get('data')
                     episodes = episodes + more_episodes
@@ -155,92 +156,95 @@ def ombi(configfile, dbfile, device, log_debug):
 
     for r in requested_movies:
         if bool(r.get("approved")):
-            tmdbid = r.get("theMovieDbId")
-            if not db.retrieve('tmdb_' + str(tmdbid)) == 'added':
-                title = mdb(configfile, dbfile, tmdbid, mdb_api, log_debug)
-                if title:
-                    best_result = search.best_result_bl(title, configfile, dbfile)
-                    print(u"Film: " + title + u" durch Ombi hinzugefügt.")
-                    if best_result:
-                        search.download_bl(best_result, device, configfile, dbfile)
-                    if english:
-                        title = r.get('title')
+            if not bool(r.get("available")):
+                tmdbid = r.get("theMovieDbId")
+                if not db.retrieve('tmdb_' + str(tmdbid)) == 'added':
+                    title = mdb(configfile, dbfile, tmdbid, mdb_api, log_debug)
+                    if title:
                         best_result = search.best_result_bl(title, configfile, dbfile)
-                        print(u"Film: " + title + u"durch Ombi hinzugefügt.")
+                        print(u"Film: " + title + u" durch Ombi hinzugefügt.")
                         if best_result:
                             search.download_bl(best_result, device, configfile, dbfile)
-                    db.store('tmdb_' + str(tmdbid), 'added')
+                        if english:
+                            title = r.get('title')
+                            best_result = search.best_result_bl(title, configfile, dbfile)
+                            print(u"Film: " + title + u"durch Ombi hinzugefügt.")
+                            if best_result:
+                                search.download_bl(best_result, device, configfile, dbfile)
+                        db.store('tmdb_' + str(tmdbid), 'added')
 
     for r in requested_shows:
         tvdbid = r.get("tvDbId")
         infos = None
-        child_request = r.get("childRequests")
-        for cr in child_request:
+        child_requests = r.get("childRequests")
+        for cr in child_requests:
             if bool(cr.get("approved")):
-                details = cr.get("seasonRequests")
-                for season in details:
-                    sn = season.get("seasonNumber")
-                    eps = []
-                    episodes = season.get("episodes")
-                    for episode in episodes:
-                        enr = episode.get("episodeNumber")
-                        s = str(sn)
-                        if len(s) == 1:
-                            s = "0" + s
-                        s = "S" + s
-                        e = str(enr)
-                        if len(e) == 1:
-                            e = "0" + e
-                        se = s + "E" + e
-                        if not db.retrieve('tvdb_' + str(tvdbid) + '_' + se) == 'added':
-                            eps.append(enr)
-                    if eps:
-                        if not infos:
-                            infos = tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug)
-                        if infos:
-                            title = infos[0]
-                            all_eps = infos[1]
-                            if all_eps:
-                                check_sn = all_eps.get(sn)
-                            else:
-                                check_sn = False
-                            if check_sn:
-                                sn_length = len(eps)
-                                check_sn_length = len(check_sn)
-                                if check_sn_length > sn_length:
-                                    for ep in eps:
-                                        e = str(ep)
-                                        if len(e) == 1:
-                                            e = "0" + e
-                                        se = s + "E" + e
+                if not bool(cr.get("available")):
+                    details = cr.get("seasonRequests")
+                    for season in details:
+                        sn = season.get("seasonNumber")
+                        eps = []
+                        episodes = season.get("episodes")
+                        for episode in episodes:
+                            if not bool(episode.get("available")):
+                                enr = episode.get("episodeNumber")
+                                s = str(sn)
+                                if len(s) == 1:
+                                    s = "0" + s
+                                s = "S" + s
+                                e = str(enr)
+                                if len(e) == 1:
+                                    e = "0" + e
+                                se = s + "E" + e
+                                if not db.retrieve('tvdb_' + str(tvdbid) + '_' + se) == 'added':
+                                    eps.append(enr)
+                        if eps:
+                            if not infos:
+                                infos = tvdb(configfile, dbfile, tvdbid, tvd_user, tvd_userkey, tvd_api, log_debug)
+                            if infos:
+                                title = infos[0]
+                                all_eps = infos[1]
+                                if all_eps:
+                                    check_sn = all_eps.get(sn)
+                                else:
+                                    check_sn = False
+                                if check_sn:
+                                    sn_length = len(eps)
+                                    check_sn_length = len(check_sn)
+                                    if check_sn_length > sn_length:
+                                        for ep in eps:
+                                            e = str(ep)
+                                            if len(e) == 1:
+                                                e = "0" + e
+                                            se = s + "E" + e
+                                            best_result = search.best_result_sj(title, configfile, dbfile)
+                                            if best_result:
+                                                add_episode = search.download_sj(best_result, se, device, configfile,
+                                                                                 dbfile)
+                                                if not add_episode:
+                                                    add_season = search.download_sj(best_result, s, device, configfile,
+                                                                                    dbfile)
+                                                    for e in eps:
+                                                        e = str(e)
+                                                        if len(e) == 1:
+                                                            e = "0" + e
+                                                        se = s + "E" + e
+                                                        db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
+                                                    if not add_season:
+                                                        log_debug(
+                                                            u"Konnte kein Release für " + title + " " + se + "finden.")
+                                                    break
+                                            db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
+                                    else:
                                         best_result = search.best_result_sj(title, configfile, dbfile)
                                         if best_result:
-                                            add_episode = search.download_sj(best_result, se, device, configfile,
-                                                                             dbfile)
-                                            if not add_episode:
-                                                add_season = search.download_sj(best_result, s, device, configfile,
-                                                                                dbfile)
-                                                for e in eps:
-                                                    e = str(e)
-                                                    if len(e) == 1:
-                                                        e = "0" + e
-                                                    se = s + "E" + e
-                                                    db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
-                                                if not add_season:
-                                                    log_debug(
-                                                        u"Konnte kein Release für " + title + " " + se + "finden.")
-                                                break
-                                        db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
-                                else:
-                                    best_result = search.best_result_sj(title, configfile, dbfile)
-                                    if best_result:
-                                        search.download_sj(best_result, s, device, configfile, dbfile)
-                                    for ep in eps:
-                                        e = str(ep)
-                                        if len(e) == 1:
-                                            e = "0" + e
-                                        se = s + "E" + e
-                                        db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
-                                print(u"Serie/Staffel/Episode: " + title + u" durch Ombi hinzugefügt.")
+                                            search.download_sj(best_result, s, device, configfile, dbfile)
+                                        for ep in eps:
+                                            e = str(ep)
+                                            if len(e) == 1:
+                                                e = "0" + e
+                                            se = s + "E" + e
+                                            db.store('tvdb_' + str(tvdbid) + '_' + se, 'added')
+                                    print(u"Serie/Staffel/Episode: " + title + u" durch Ombi hinzugefügt.")
 
     return device
